#230 Singularize Artwork Test Stubbing
